Gif rendering (and it mostly works)
Doesn't work with all gifs.
Most gifs show a distorted frame every x frames.
A few gifs don't load due to being too big.

+++ src/renderer/gif2d_pipeline.rs
@@ -17,7 +17,7 @@ pub struct Gif2dInstanceRaw {
     pub pos: [f32; 2],
     pub size: [f32; 2],
     pub color: [f32; 4],
-    pub frame: u32,
+    pub frame: f32,
     pub padding: [f32; 3],
 }
